add common setup function
