+++ packages/cognito/src/token.ts
@@ -11,9 +11,8 @@ import axios from "axios"
 import {parse, ParsedUrlQuery, stringify} from "querystring"
 
 import {PrivateKey} from "jsonwebtoken"
-import jwksClient from "jwks-rsa"
 
-import {verifyIdToken, OidcConfig} from "@cpt-ui-common/authFunctions"
+import {verifyIdToken, initializeOidcConfig} from "@cpt-ui-common/authFunctions"
 import {MiddyErrorHandler} from "@cpt-ui-common/middyErrorHandler"
 
 import {formatHeaders, rewriteBodyToAddSignedJWT} from "./helpers"
@@ -46,43 +45,10 @@ MOCK_IDP_TOKEN_PATH
 
 const logger = new Logger({serviceName: "token"})
 const useMock: boolean = process.env["useMock"] === "true"
-// Create a JWKS client for cis2 and mock
-// this is outside functions so it can be re-used
-const cis2JwksUri = process.env["CIS2_OIDCJWKS_ENDPOINT"] as string
-const cis2JwksClient = jwksClient({
-  jwksUri: cis2JwksUri,
-  cache: true,
-  cacheMaxEntries: 5,
-  cacheMaxAge: 3600000 // 1 hour
-})
 
-const cis2OidcConfig: OidcConfig = {
-  oidcIssuer: process.env["CIS2_OIDC_ISSUER"] ?? "",
-  oidcClientID: process.env["CIS2_OIDC_CLIENT_ID"] ?? "",
-  oidcJwksEndpoint: process.env["CIS2_OIDCJWKS_ENDPOINT"] ?? "",
-  oidcUserInfoEndpoint: process.env["CIS2_USER_INFO_ENDPOINT"] ?? "",
-  userPoolIdp: process.env["CIS2_USER_POOL_IDP"] ?? "",
-  jwksClient: cis2JwksClient,
-  tokenMappingTableName: process.env["TokenMappingTableName"] ?? ""
-}
-
-const mockJwksUri = process.env["MOCK_OIDCJWKS_ENDPOINT"] as string
-const mockJwksClient = jwksClient({
-  jwksUri: mockJwksUri,
-  cache: true,
-  cacheMaxEntries: 5,
-  cacheMaxAge: 3600000 // 1 hour
-})
-
-const mockOidcConfig: OidcConfig = {
-  oidcIssuer: process.env["MOCK_OIDC_ISSUER"] ?? "",
-  oidcClientID: process.env["MOCK_OIDC_CLIENT_ID"] ?? "",
-  oidcJwksEndpoint: process.env["MOCK_OIDCJWKS_ENDPOINT"] ?? "",
-  oidcUserInfoEndpoint: process.env["MOCK_USER_INFO_ENDPOINT"] ?? "",
-  userPoolIdp: process.env["MOCK_USER_POOL_IDP"] ?? "",
-  jwksClient: mockJwksClient,
-  tokenMappingTableName: process.env["TokenMappingTableName"] ?? ""
-}
+// Create a config for cis2 and mock
+// this is outside functions so it can be re-used and caching works
+const {mockOidcConfig, cis2OidcConfig} = initializeOidcConfig()
 
 const TokenMappingTableName = process.env["TokenMappingTableName"] as string
 const jwtPrivateKeyArn = process.env["jwtPrivateKeyArn"] as string
