+++ packages/trackerUserInfoLambda/src/handler.ts
@@ -5,9 +5,8 @@ import {DynamoDBClient} from "@aws-sdk/client-dynamodb"
 import {DynamoDBDocumentClient} from "@aws-sdk/lib-dynamodb"
 import middy from "@middy/core"
 import inputOutputLogger from "@middy/input-output-logger"
-import jwksClient from "jwks-rsa"
 import {MiddyErrorHandler} from "@cpt-ui-common/middyErrorHandler"
-import {getUsernameFromEvent, fetchAndVerifyCIS2Tokens, OidcConfig} from "@cpt-ui-common/authFunctions"
+import {getUsernameFromEvent, fetchAndVerifyCIS2Tokens, initializeOidcConfig} from "@cpt-ui-common/authFunctions"
 import {fetchUserInfo, updateDynamoTable} from "./userInfoHelpers"
 
 /*
@@ -35,43 +34,9 @@ const logger = new Logger({serviceName: "trackerUserInfo"})
 const dynamoClient = new DynamoDBClient({})
 const documentClient = DynamoDBDocumentClient.from(dynamoClient)
 
-// Create a JWKS client for cis2 and mock
-// this is outside functions so it can be re-used
-const cis2JwksUri = process.env["CIS2_OIDCJWKS_ENDPOINT"] as string
-const cis2JwksClient = jwksClient({
-  jwksUri: cis2JwksUri,
-  cache: true,
-  cacheMaxEntries: 5,
-  cacheMaxAge: 3600000 // 1 hour
-})
-
-const cis2OidcConfig: OidcConfig = {
-  oidcIssuer: process.env["CIS2_OIDC_ISSUER"] ?? "",
-  oidcClientID: process.env["CIS2_OIDC_CLIENT_ID"] ?? "",
-  oidcJwksEndpoint: process.env["CIS2_OIDCJWKS_ENDPOINT"] ?? "",
-  oidcUserInfoEndpoint: process.env["CIS2_USER_INFO_ENDPOINT"] ?? "",
-  userPoolIdp: process.env["CIS2_USER_POOL_IDP"] ?? "",
-  jwksClient: cis2JwksClient,
-  tokenMappingTableName: process.env["TokenMappingTableName"] ?? ""
-}
-
-const mockJwksUri = process.env["MOCK_OIDCJWKS_ENDPOINT"] as string
-const mockJwksClient = jwksClient({
-  jwksUri: mockJwksUri,
-  cache: true,
-  cacheMaxEntries: 5,
-  cacheMaxAge: 3600000 // 1 hour
-})
-
-const mockOidcConfig: OidcConfig = {
-  oidcIssuer: process.env["MOCK_OIDC_ISSUER"] ?? "",
-  oidcClientID: process.env["MOCK_OIDC_CLIENT_ID"] ?? "",
-  oidcJwksEndpoint: process.env["MOCK_OIDCJWKS_ENDPOINT"] ?? "",
-  oidcUserInfoEndpoint: process.env["MOCK_USER_INFO_ENDPOINT"] ?? "",
-  userPoolIdp: process.env["MOCK_USER_POOL_IDP"] ?? "",
-  jwksClient: mockJwksClient,
-  tokenMappingTableName: process.env["TokenMappingTableName"] ?? ""
-}
+// Create a config for cis2 and mock
+// this is outside functions so it can be re-used and caching works
+const {mockOidcConfig, cis2OidcConfig} = initializeOidcConfig()
 
 const MOCK_MODE_ENABLED = process.env["MOCK_MODE_ENABLED"]
 const tokenMappingTableName = process.env["TokenMappingTableName"] ?? ""
